requirements.txt
deep_translator==1.7.0
nltk==3.7
python-dotenv==0.19.2
tweepy==4.6.0
